Fix unit 3 - async didn't work correctly - removed nested functions
