+++ nodejs-files/index-unit-3.js
@@ -1,39 +1,25 @@
+const path = require("path");
 const fs = require("fs").promises;
 
-async function findSalesFiles(folderName) {
-  // this array will hold sales files as they are found
-  let salesFiles = [];
+async function findFiles(folderName) {
 
-  async function findFiles(folderName) {
-    // read all the items in the current folder
-    const items = await fs.readdir(folderName, { withFileTypes: true });
+    let results = []
 
-    // iterate over each found item
-    for (item of items) {
-      // if the item is a directory, it will need to be searched for files
-      if (item.isDirectory()) {
-        // search this directory for files (this is recursion!)
-        await findFiles(`${folderName}/${item.name}`);
-      } else {
-        // Make sure the discovered file is a sales.json file
-        if (item.name === "sales.json") {
-          // store the file path in the salesFiles array
-          await salesFiles.push(`${folderName}/${item.name}`);
+    results.push(`${folderName}`);
+
+    const relativePathToFolder = path.join(__dirname, folderName);
+    const items = await fs.readdir(relativePathToFolder, { withFileTypes: true });
+
+    for (const item of items) {
+        if (item.isDirectory()) {
+            const resultsReturned = await findFiles(`${folderName}/${item.name}`);
+            results = results.concat(resultsReturned);
+        } else {
+            results.push(`${folderName}/${item.name}`);
         }
-      }
     }
-  }
 
-  // find the sales files
-  await findFiles(folderName);
-
-  // return the array of found file paths
-  return salesFiles;
+    return results;
 }
 
-async function main() {
-  const salesFiles = await findSalesFiles("stores");
-  console.log(salesFiles);
-}
-
-main();
+findFiles("stores").then((results) => console.log(results));
